+++ CHANGELOG.md
@@ -8,6 +8,7 @@ O versionamento deste projeto é aderente aos princípios de [Semantic Versionin
 
 ### Added
 
+- Criação do utilitário `<cps-format-bytes>`.
 - Criação do utilitário `<cps-format-date>`.
 - Criação do utilitário `<cps-format-number>`.